feat: Upgrades to React Native 0.68.2, adds rename CLI command (#1950 by @jamonholmgren)
This version upgrades the boilerplate to the latest version of React Native (0.68.2), as well as multiple other dependencies.

### Notable boilerplate changes:

* Upgrades to React Native 0.68.2, allowing for the (optional) new architecture
* .npmrc file added to avoid pnpm peer dependency issues
* Fixes recursive type issue in i18n files
* Fixes some remaining symlink issues

### Notable other changes:

* Adds new `ignite-cli rename` feature and removes dependency on `react-native-rename`. More below!
* Added check for leftover "HelloWorld" files in CI tests to protect against rename issues
* CI has been updated to a newer image

#### Rename

We've been relying on the package `react-native-rename` for a few versions now, but it doesn't seem to be keeping pace with the changes in the React Native world. So, I've built my own version of this. It runs automatically when you spin up a vanilla Ignite project (non-Expo), but also can be run on any React Native project!

Just run:

```
npx ignite-cli rename
```

It'll ask for your new app name and bundle identifier and make the required changes.

If you run into problems, please report them!